refactor(frontend): group related action types  & reducers
