--- redux/constants/signIn.js
@@ -1,3 +1,0 @@
-export const SIGNIN_REQUEST = 'SIGNIN_REQUEST';
-export const SIGNIN_SUCCESS = 'SIGNIN_SUCCESS';
-export const SIGNIN_ERROR = 'SIGNIN_ERROR';
